Fixes a compilation error in `IronPunch.java`.
The `getNearbyEntities` method in `Utils.java` was expecting an integer for the radius, but a double was passed. The method has been updated to accept a double to make it more flexible and to fix the bug.

+++ src/main/java/com/jules/kitpvp/util/Utils.java
@@ -72,7 +72,7 @@ public class Utils {
         return (double) tmp / factor;
     }
 
-    public static List<Entity> getNearbyEntities(Location l, int size) {
+    public static List<Entity> getNearbyEntities(Location l, double size) {
         List<Entity> entities = new ArrayList<Entity>();
         for(Entity ent : l.getWorld().getEntities()) {
             if(!(ent instanceof LivingEntity)) continue;
